requirements.txt
pandas>=2.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
openpyxl>=3.1.0  # For Excel file support